Refine waitlist fulfillment chain
