+++ services/waitlist/src/adapters/mod.rs
@@ -1,0 +1,3 @@
+pub mod http;
+pub mod messaging;
+pub mod storage;
